requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-html==4.1.1
requests==2.31.0
httpx==0.25.2
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99
protobuf==4.25.1
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
tabulate==0.9.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1
ipython==8.18.1